requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-index-vector-stores-weaviate
llama-index-llms-openai
llama-index-core
weaviate-client
openai
python-dotenv==1.0.0
gunicorn==22.0.0
llama-index-embeddings-openai
llama-index 